// src/types/types.ts
export type FindGroupsListType = "nearby" | "recommended" | "personal" | "filter";